Cambios realizados vinculados al editar la captacion: 1- Modificacion de la conts tabla, agregando las dos acciones en el datatable (editar y eliminar). 2- Se agrego la class updateInstanciaView en el archivo view de la app seccion. 3- Creacion de la template crear instancia. 3- Modificacion del apartado ---MODIFICACIONES DE LAS COSAS CAPTADAS DEL CUESTIONARIO--- en el archivo funciones.js, especificamente en ---INSTANCIA--. 4- Se agrego en el archivo view de la app seccion la class eliminarInstanciaView.

+++ apps/seccion/templatetags/filters.py
@@ -86,7 +86,7 @@ def getInfoVerificacion(cuestionarioId, seccion):
     try:
         query = verificacion.objects.get(cuestionario_fk__id=cuestionarioId, seccion_id__nombre=seccion)
         noCoinciden = query.indicadoresVerificados - query.indicadoresCoinciden
-        porciento = noCoinciden*100/query.indicadoresVerificados
+        porciento = int(noCoinciden*100/query.indicadoresVerificados)
         return porciento
     except:
         return None
